chore: reorganize workflow ops (#1850)
