+++ packages/client/hmi-client/src/workflow/ops/calibrate-ensemble-ciemss/tera-calibrate-ensemble-ciemss.vue
@@ -227,7 +227,7 @@ import ChartDataLabels from 'chartjs-plugin-datalabels';
 import { ChartConfig, RunResults } from '@/types/SimulateConfig';
 import { IProject } from '@/types/Project';
 import { setupDatasetInput } from '@/services/calibrate-workflow';
-import TeraSimulateChart from './tera-simulate-chart.vue';
+import TeraSimulateChart from '@/workflow/tera-simulate-chart.vue';
 import {
 	CalibrateEnsembleCiemssOperationState,
 	EnsembleCalibrateExtraCiemss
