+++ packages/client/hmi-client/src/workflow/ops/dataset-transformer/mod.ts
@@ -1,0 +1,3 @@
+export * from './dataset-transformer-operation';
+export { default as TeraDatasetTransformer } from './tera-dataset-transformer.vue';
+export { default as TeraDatasetTransformerNode } from './tera-dataset-transformer-node.vue';
