+++ packages/client/hmi-client/src/workflow/ops/simulate-ciemss/tera-simulate-ciemss.vue
@@ -1,7 +1,7 @@
 <template>
 	<section class="tera-simulate">
 		<div class="simulate-header">
-			<span class="simulate-header-label">Simulate (deterministic)</span>
+			<span class="simulate-header-label">Simulate (probabilistic)</span>
 			<div class="simulate-header p-buttonset">
 				<Button
 					label="Input"
@@ -12,7 +12,7 @@
 					@click="activeTab = SimulateTabs.input"
 				/>
 				<Button
-					label="Ouput"
+					label="Output"
 					severity="secondary"
 					icon="pi pi-sign-out"
 					size="small"
@@ -25,13 +25,53 @@
 			v-if="activeTab === SimulateTabs.output && node?.outputs.length"
 			class="simulate-container"
 		>
+			<div class="datatable-header">
+				<div class="datatable-header-title">
+					{{ `${rawDataKeys.length} columns | ${parsedRawData.length} rows` }}
+				</div>
+				<div class="datatable-header-select-container">
+					<MultiSelect v-model="selectedCols" :options="rawDataKeys" placeholder="Select columns" />
+				</div>
+			</div>
+			<div
+				class="p-datatable p-component p-datatable-scrollable p-datatable-responsive-scroll p-datatable-gridlines p-datatable-grouped-header"
+			>
+				<div class="p-datatable-wrapper">
+					<table class="p-datatable-table p-datatable-scrollable-table editable-cells-table">
+						<thead class="p-datatable-thead">
+							<tr>
+								<th
+									v-for="(header, i) of selectedCols.length ? selectedCols : rawDataKeys"
+									:key="i"
+									class="p-frozen-column"
+								>
+									{{ header }}
+								</th>
+							</tr>
+						</thead>
+						<tbody class="p-datatable-tbody">
+							<tr v-for="(data, i) of rawDataRenderedRows" :key="i">
+								<td v-for="(key, j) of selectedCols.length ? selectedCols : rawDataKeys" :key="j">
+									{{ data[key] }}
+								</td>
+							</tr>
+						</tbody>
+					</table>
+				</div>
+			</div>
+			<Paginator
+				v-model:first="paginatorFirst"
+				v-model:rows="paginatorRows"
+				:totalRecords="parsedRawData.length"
+				:rowsPerPageOptions="[5, 10, 20, 50]"
+			/>
 			<tera-simulate-chart
 				v-for="(cfg, index) of node.state.chartConfigs"
 				:key="index"
 				:run-results="runResults"
 				:chartConfig="cfg"
+				has-mean-line
 				@configuration-change="configurationChange(index, $event)"
-				color-by-run
 			/>
 			<Button
 				class="add-chart"
@@ -71,8 +111,17 @@
 			<div class="simulate-model">
 				<Accordion :multiple="true" :active-index="[0, 1, 2]">
 					<AccordionTab>
-						<template #header> {{ modelConfiguration?.configuration.name }} </template>
-						<model-diagram v-if="model" :model="model" :is-editable="false" />
+						<template #header> Model </template>
+						<tera-model-diagram v-if="model" :model="model" :is-editable="false" />
+					</AccordionTab>
+					<AccordionTab>
+						<template #header> Model configuration </template>
+						<tera-model-configurations
+							v-if="model"
+							:model="model"
+							:model-configurations="modelConfigurations"
+							:feature-config="{ isPreview: true }"
+						/>
 					</AccordionTab>
 					<AccordionTab>
 						<template #header> Simulation time range </template>
@@ -97,6 +146,21 @@
 							</div>
 						</div>
 					</AccordionTab>
+					<AccordionTab>
+						<template #header> Other options </template>
+						<div class="sim-tspan-container">
+							<div class="sim-tspan-group">
+								<label for="4">Number of stochastic samples</label>
+								<InputNumber
+									id="4"
+									class="p-inputtext-sm"
+									v-model="numSamples"
+									inputId="integeronly"
+									:min="2"
+								/>
+							</div>
+						</div>
+					</AccordionTab>
 				</Accordion>
 			</div>
 		</div>
@@ -105,34 +169,37 @@
 
 <script setup lang="ts">
 import _ from 'lodash';
-import { ref, onMounted, computed } from 'vue';
+import { ref, onMounted, computed, watch } from 'vue';
 import Accordion from 'primevue/accordion';
 import AccordionTab from 'primevue/accordiontab';
+import MultiSelect from 'primevue/multiselect';
 import Button from 'primevue/button';
 import InputNumber from 'primevue/inputnumber';
-import { ModelConfiguration, Model, TimeSpan } from '@/types/Types';
+import Paginator from 'primevue/paginator';
+import { Model, TimeSpan, ModelConfiguration } from '@/types/Types';
 import { ChartConfig, RunResults } from '@/types/SimulateConfig';
-
+import { getModel, getModelConfigurations } from '@/services/model';
 import { getModelConfigurationById } from '@/services/model-configurations';
-import ModelDiagram from '@/components/model/petrinet/model-diagrams/tera-model-diagram.vue';
-
-import { getSimulation, getRunResult } from '@/services/models/simulation-service';
-import { getModel } from '@/services/model';
-import { saveDataset } from '@/services/dataset';
-import { csvParse } from 'd3';
+import { getRunResultCiemss } from '@/services/models/simulation-service';
+import TeraModelDiagram from '@/components/model/petrinet/model-diagrams/tera-model-diagram.vue';
+import TeraModelConfigurations from '@/components/model/petrinet/tera-model-configurations.vue';
+import TeraSimulateChart from '@/workflow/tera-simulate-chart.vue';
 import { WorkflowNode } from '@/types/workflow';
 import { workflowEventBus } from '@/services/workflow';
 import { IProject } from '@/types/Project';
+import { saveDataset } from '@/services/dataset';
 import InputText from 'primevue/inputtext';
-import { SimulateJuliaOperationState } from './simulate-julia-operation';
-import TeraSimulateChart from './tera-simulate-chart.vue';
+import { SimulateCiemssOperationState } from './simulate-ciemss-operation';
 
 const props = defineProps<{
 	node: WorkflowNode;
 	project?: IProject;
 }>();
 
+const hasValidDatasetName = computed<boolean>(() => saveAsName.value !== '');
+
 const timespan = ref<TimeSpan>(props.node.state.currentTimespan);
+const numSamples = ref<number>(props.node.state.numSamples);
 
 enum SimulateTabs {
 	input,
@@ -142,15 +209,19 @@ enum SimulateTabs {
 const activeTab = ref(SimulateTabs.input);
 
 const model = ref<Model | null>(null);
+const modelConfigurations = ref<ModelConfiguration[]>([]);
+const parsedRawData = ref<any>();
+const runConfigs = ref<any>({});
 const runResults = ref<RunResults>({});
-const modelConfiguration = ref<ModelConfiguration | null>(null);
-const completedRunId = computed<string | undefined>(() => props?.node?.outputs?.[0]?.value?.[0]);
-const hasValidDatasetName = computed<boolean>(() => saveAsName.value !== '');
 const showSaveInput = ref(<boolean>false);
 const saveAsName = ref(<string | null>'');
+const selectedCols = ref<string[]>([]);
+const paginatorRows = ref(10);
+const paginatorFirst = ref(0);
+const completedRunId = computed<string | undefined>(() => props?.node?.outputs?.[0]?.value?.[0]);
 
 const configurationChange = (index: number, config: ChartConfig) => {
-	const state: SimulateJuliaOperationState = _.cloneDeep(props.node.state);
+	const state: SimulateCiemssOperationState = _.cloneDeep(props.node.state);
 	state.chartConfigs[index] = config;
 
 	workflowEventBus.emitNodeStateChange({
@@ -161,7 +232,7 @@ const configurationChange = (index: number, config: ChartConfig) => {
 };
 
 const addChart = () => {
-	const state: SimulateJuliaOperationState = _.cloneDeep(props.node.state);
+	const state: SimulateCiemssOperationState = _.cloneDeep(props.node.state);
 	state.chartConfigs.push(_.last(state.chartConfigs) as ChartConfig);
 
 	workflowEventBus.emitNodeStateChange({
@@ -183,35 +254,37 @@ onMounted(async () => {
 	if (!port.value) return;
 	const simulationId = port.value[0];
 
-	const simulationObj = await getSimulation(simulationId as string);
-	if (!simulationObj) return;
+	const modelConfigId = nodeObj.inputs[0].value?.[0];
+	const modelConfiguration = await getModelConfigurationById(modelConfigId);
+	if (modelConfiguration) {
+		model.value = await getModel(modelConfiguration.modelId);
+		if (model.value) modelConfigurations.value = await getModelConfigurations(model.value.id);
+	}
 
-	const executionPayload = simulationObj.executionPayload;
-
-	if (!executionPayload) return;
-
-	const modelConfigurationId = (simulationObj.executionPayload as any).model_config_id;
-	const modelConfigurationObj = await getModelConfigurationById(modelConfigurationId);
-	const modelId = modelConfigurationObj.modelId;
-	model.value = await getModel(modelId);
-
-	// Fetch run results
-	await Promise.all(
-		port.value.map(async (runId) => {
-			const resultCsv = await getRunResult(runId, 'result.csv');
-			const csvData = csvParse(resultCsv);
-			if (modelConfigurationObj) {
-				const parameters = modelConfigurationObj.configuration.semantics.ode.parameters;
-				csvData.forEach((row) =>
-					parameters.forEach((parameter) => {
-						row[parameter.id] = parameter.value;
-					})
-				);
-			}
-			runResults.value[runId] = csvData as any;
-		})
-	);
+	const output = await getRunResultCiemss(simulationId);
+	parsedRawData.value = output.parsedRawData;
+	runResults.value = output.runResults;
+	runConfigs.value = output.runConfigs;
 });
+
+watch(
+	() => numSamples.value,
+	(n) => {
+		const state: SimulateCiemssOperationState = _.cloneDeep(props.node.state);
+		state.numSamples = n;
+		workflowEventBus.emitNodeStateChange({
+			workflowId: props.node.workflowId,
+			nodeId: props.node.id,
+			state
+		});
+	}
+);
+
+const rawDataKeys = computed(() => Object.keys(parsedRawData.value[0]));
+
+const rawDataRenderedRows = computed(() =>
+	parsedRawData.value.slice(paginatorFirst.value, paginatorFirst.value + paginatorRows.value)
+);
 </script>
 
 <style scoped>
@@ -244,7 +317,7 @@ onMounted(async () => {
 }
 
 .simulate-chart {
-	margin: 2em 1em;
+	margin: 2em 1.5em;
 }
 
 .sim-tspan-container {
@@ -262,5 +335,22 @@ onMounted(async () => {
 ::v-deep .p-inputnumber-input,
 .p-inputwrapper {
 	width: 100%;
+}
+
+.datatable-header {
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	margin: 0.5em;
+	position: relative;
+}
+
+.datatable-header-select-container {
+	min-width: 0;
+}
+
+.datatable-header-title {
+	white-space: nowrap;
+	margin-right: 1em;
 }
 </style>
