+++ packages/client/hmi-client/src/workflow/ops/stratify-julia/tera-stratify.vue
@@ -221,7 +221,7 @@ import { addAsset } from '@/services/project';
 import { stratify } from '@/model-representation/petrinet/petrinet-service';
 import useResourcesStore from '@/stores/resources';
 import { workflowEventBus } from '@/services/workflow';
-import { ModelOperation } from '@/components/workflow/model-operation';
+import { ModelOperation } from '@/workflow/ops/model/model-operation';
 import TeraStrataModelDiagram from '@/components/model/petrinet/model-diagrams/tera-strata-model-diagram.vue';
 import TeraTypedModelDiagram from '@/components/model/petrinet/model-diagrams/tera-typed-model-diagram.vue';
 import TeraStratifyOutputModelDiagram from '@/components/model/petrinet/model-diagrams/tera-stratify-output-model-diagram.vue';
