+++ packages/client/hmi-client/src/workflow/tera-simulate-chart.vue
